backend: Serve frontend for all unhandled routes
We only serve our frontend build when the server is requested for "/"
route, this is fine as when we go to the home page and refresh, the
server serves the correct files and then every route we visit is handled
by frontend routing. But the problem now comes when we visit some
frontend route and press refresh button,, now our app asks server
please serve the following routes file but the server doesn't have such
a route as that route was created by frontend and backend doesn't have
any idea about it and thus server sends a 404 error. Thus this patch
adds a NoRoute which serves frontend when any unhandled route is
requested at the backend.