Don't fetch the instances stats when fetching groups
The query that fetched the groups was getting the instances stats for
every group as it was very convenient to access them directly in the
resulting group object. However, this meant that every time the backend
or frontend needed to get a group's information, extra queries would
also be run to get the stats and this was slowing down every query which
could become a problem when e.g. multiple FCL instances try to get
update information from Nebraska.

In order to fix that, this patch decouples the instances stats query
from the simple get-group one (skipping that extra processing).
Of course, in order to make up for that, it also adds direct ways for
getting a group's instances stats as well as the required changes for
the frontend to keep running.