backend: Create release process for build
We added a version of nebraska as a string to the code in a previous pr,
so now when doing a release, we will need to remember to bump it, but
this requires manual work and sometime we might even forget to do this
manual version change. Thus this patch creates a release process for
build using ldflag that adds version from tag to build binary.