backend: Limit the database connections
The database connections were not getting limited, so theoretically
we could end up with unlimited connections to the database, which
increase the CPU usage and end up alowing down everything.

To fix the issue mentioned above, this patch adds a way to limit the
database maximum open/idle connections, as well as the connections'
maximum lifetime. By default the maximum open/idle now are 25, and
their lifetime is 5 minutes.