backend: Add duration query param handling capability.
We didn't have the functionality to fetch data for the
versionCountTimeline and statusCountTimeline with a custom duration,
i.e. we were always fetching data for a 1 day period. This patch changes
the routes for both chart getter functions to take in a duration field
and adds logic for changing the interval based on that duration.