frontend: Add the period param to the instances list
Users usually visit the instances list page from the update progress
chart but we were not passing the period so that we can fetch the
instances based on it. This resulted in, e.g. users seeing a number of
instances for the last week, and then seeing a completely different
number/list when visiting the instances list.

Therefore this patch adds the period query parameter when visiting
the instance list view from the group extended view and changes the
logic to fetch instances based on that.