frontend: Add period selector logic for instances charts
The VersionBreakdownChart and StatusBreakdownChart needed a way to
select the period for which they refer to and we should cache the
instances data so a user doesn't have to wait multiple times for the
likely very similar data. Thus this patch adds a store for
caching/fetching the instances and changes the logic in the mentioned
charts to use the period/duration parameter.