frontend: Handle race condition for instances' period selection
When we switch between duration links there is a race between the API
calls for switched periods and thus this leads to inconsistency into
what data will be shown. This patch fixes this problem by cancelling
a previous fetch action that may be on-going (by not using the results
when the fetch call finishes).