frontend: Fix channel color overriding on edit
When we edit a channel it seems the next update on some other channel
was getting the previous one's color. This was happening because we were
not updating the color in the edit dialog when the data changed.

This patch fixes this issue by updating the color state whenever the
data prop changes.