frontend: Change ticks for a month period in the instances' charts
Earlier for 1 month period we were showing only four ticks but those
were not very helpful for understanding the chart when the 30 days
period was selected.

Thus, this patch changes the logic for tick generation if the month
period is selected, in order to show ticks over a interval of 6 days.