frontend: Show periods that have no instances available in gray
We were showing the duration links in the group extended view even if
they didn't have any instances to report, so the user had to click every
period link in order to find out if there were instances available or
not (which is not a great UX).

In order to improve this situation this patch sets the color of periods
with no instances reported in them as gray. While the links are still
clickable (so users can go into the view and see the message clarifying
that there are no instances), this solution gives a hint to the user
about what the interesting periods are.