frontend: Simplify EmptyContent
Instead of checking the EmptyContent's children types and wrap them in
a Typography element or add them directly, we should just wrap all the
children in a Typography element, since the contents of EmptyContent
are supposed to be textual anyway.

Since these changes break the tests, an update to those tests was
written by Ashu Ghildiyal and is included here.