frontend: Fix charts click and duration change bug
When we click on the instances' charts (Status and Version Breakdown)
and then change the duration, there was a crash caused because our data
didn't have a selected field on it, and thus asking for timestamp from
it was resulting in the error.

This patch fixes by adding a check before accessing the related data
value.