backend: Replace dat package
We were using the dat package for querying the database but it is no
longer maintained and we thus needed to switch to another alternative
which is healthy and well maintained. So this patch replaces dat by
using pgx for the PostgresSQL driver, goqu for abstracting the SQL
queries, and the null package for nullable types.

Since a couple of things changed in the way the queries respond, some
tests are also updated.