backend: Grant update only after checking the version
When the fast path was added to re-grant an update that was already
granted (6bcd6fdbfd2d7ad29b17ab830219ed1aad34be1a) this skipped
the version check, resulting in granting an update to clients that
already run this version. The DB state of granting an update should
not be more important than the version of the instance because the
DB can either lack entries (due to the client not sending them or
the server dropping them) or it can also be tricked by a malicious
third party pretending to be the client but with a lower app version.
Only re-grant an update after performing all checks that are normally
done as well, i.e., if the client's version needs an update and if
the package wasn't forbidden for this channel in the mean time. Set
the instance status to InstanceStatusComplete to keep the right entry
in the log and also don't allow the instance to bypass the rollout
policy for a future update.