backend: Add limit to request size for /omaha and /v1/update requests
Nebraska by design has two unauthenticated API endpoints to handle
instance update requests. Both are handled by the Omaha request handler
that has to accept data from any clients. So we should limit the size of
the allowed requests' data, in order to prevent eventual abuse that may
lead to extra load on the server.

Thus, this patch adds a limit to the request body size for /omaha and
/v1/update.