backend: Improve the use of errors for requests that are too large
We want to improve how the request handler returns errors for the
case of when the request's body is too large. So these changes
return a BadRequest http error in such cases.
This patch improves the logging for omaha request failure and also
changes how we respond to large size request to omaha.