backend: Fix type for InstanceStats
We were having int type for values in InstanceStats struct that has the
possiblity of being null or int and thus when we were Scanning over
InstanceStats with null values it was giving us scan error. Thus this
patch updates the type from int to null.Int(meaning it can be null or
int). This patch also updates the tests to compare correct type values
for InstanceStats properties in updates_test.go.