frontend: Make the group's updates policy value a string
The mentioned value, in the group extended view was just the value node
in JSX and not correctly broken in lines, so the result was that the
number of updates allowed didn't have a space between it and the rest of
the text. This patch fixes it by making that line a string.