build: Fix a cleanup bug
The snippet that sets up the postgres database in Makefile had a bug -
if creating a database or setting the timezone failed then the
container_id file wouldn't be deleted, because those two commands were
being executed in a different shells, without the trap. There were
missing backslashes to join the commands into a single shell
invocation. This would result in a subsequent `make
check-backend-with-container` to talk to a not-fully-set-up postgres
database.