tools: Add a script for running a local postgres db
Since we dropped the nebraska/postgres Dockerfile, there was no
straightforward way of running a local postgres database to run the
frontend locally. There is a bash snippet in the Makefile that is
setting up a database for unit tests. The new script reuses the
snippet, but it's refactored to make it a bit more general. See the
documentation comments in the script to see how to use it.

The script is meant to be used for testing and development only.