backend: Fix upsert query for flatcar action
Earlier we were fetching flatcar action by package id and if it didn't
exist, then we were using an insert query, otherwise an update query was
used. This process can be made much simpler by changing this to an
upsert query from goqu which is more readable, atomic, and less complex.