backend: Fix breakdown queries
We were using lateral join which for every row in table 1 runs the
defined subquery. However with large CTEs in breakdown query the lateral
join was getting very slow. The fix for this is to replace lateral joins
with simple joins so that we don't execute the subquery for each row.
Also, for the status version count timeline the query time can be
reduced by selecting only the data which falls in the required time
frame, for example if we want to query for 1 day, all data from
instance_status_history before 1 day is useless to us and that's why
this patch also adds a filter while querying from
instance_status_history (in the status version count timeline).