backend: Make filter_status_history CTE
For more readability, getting rows from instance_status_history
in status count timeline can be replaced with a seperate CTE. With the
CTE added this patch also fixes the condition of left join from a
between clause to ">= and <" expression. We want to consider
strictly a "lesser than" values, otherwise we get duplicate values at
the full hour.