backend: Optimize getApps and remove packages property from App struct
We were fetching packages alongside apps which was leading to an
increase load time for getApps queries, we already have endpoint for
getting packages which can be used to fetch packages for a particular
application. Also as we now don't pass packages with application, so
it makes sense to remove the packages property from the Application
struct.