frontend: Fix getting the packages when editing a channel
The list of available packages used to be set directly in the
Application object, but due to a recent optimization at the DB queries'
level, that's no longer the case and we need to fetch the packages
directly, which had been forgotten in the channels' List component. As
a result the channel edition was broken.

This patch fetches the packages directly, which fixes the issue
mentioned above.