frontend: Clarify the timeline charts' default legend shown
The status and version breakdown timelines charts show the legend for
the last time period in the chart, but the message above the legend
said "Showing data for the last 24 hours (click the chart to choose a
different time point)". This may lead users to think we're showing an
overall count of the instances in that period, and that would be nice,
but we don't have that functionality yet.

So in order to fix this UX issue, this patch changes the labels to make
it clear that the data shown in the legend is for the last time point of
the chart.