frontend: Delay showing the timeline charts' default legend label
The mentioned label, which explains what the legend is showing the data
for (and explains the user can click the chart) was shown from the
beginning even when the chart is loading. This doesn't look good since
the user reads a message for something that is not yet available.

This patch shows the mentioned label only when the data is finally
available.