frontend: Delay the empty message until the instances charts are loaded
The instances charts may take a second or two to load, and the empty
message was being shown in the meanwhile which could be
alarming/misleading for the users.

This patch fixes that by introducing the notion of a "loading" state
when the data is null and not empty, and shows a loader until the data
is gathered.