frontend: Fix how the color is set on ModalButton
We were earlier passing the custom hex color code to the color prop for
the AddIcon component which is incorrect as the color only takes some
specific value of already defined color class names. This patch adds the
color to the icon by using the useStyles approach and directly setting
color on the root element.