frontend: Disable typography wrapper for secondary text
We don't want the secondary text for the ListItemText component to be
wrapped in a Typography component as that will lead to dom violation
because than we can only use inline elements inside the secondary text.
This patch adds the disableTypography prop which disables wrapping the
secondary text with a Typography.