backend: Pass the error code with instance status history
When an instance reports an error, the error has a specific code. But we
right now just show a plain hard coded string without specifying the
exact error. This patch allows passing the error codes to the instance
status history list it does it so by fetching the most recent event and
getting the error code from there.