backend: Use the correct interval when fetching the status timeline
We right now were fetching status timeline for 1 hour less than the
generated timestamps which is correct in case of 1 day duration but not
when we have duration as 7 days or 30 days because for these durations
the interval is different and we want to look into that entire interval
window. This patch uses the set intervals for different durations while
fetching the status timeline.