backend: Make track name to group mapping dynamic
The mapping of Omaha track to a group ID was hardcoded for the stable,
beta, alpha, and edge channel using only the initial the group IDs.
When the groups were recreated resolving the track was broken and new
groups always had to use the group ID as track which was confusing.
Calculate the mapping dynamically based on the channel architecture and
the track name as the channel name which is also what a user would
expect. Instead of having a complex query using one of the track
name or the ID, or a two queries with the first query resolving a
possible track name and a second query using it or the ID, the
implementation uses a hash map cache which is swapped out when a group
entry needs an update. A proper solution would be to add a new slug
field for the track name and a migration to set it up for the initial
Flatcar groups. For new groups it would have to be entered in the UI
and enforced to be unique.