pkg/syncer/syncer: Add parameters for sync interval and source URL
The interval was hardcoded to one hour which is inconvenient when new
settings should be quickly propagated during a test phase. The sync
URL was also hardcoded and required source code changes when a different
source URL should has to be used.
Expose both values as parameters with the current values as default.