pkg/syncer/syncer: Reduce waiting time between two channels
The syncer was sleeping for a minute before the next channel was
synced. This makes it look like a bug and even if known it is not
really useful but rather disturbing to wait 8 minutes until all
channels are synced.
Wait for 5 seconds which is a bit waiting time to not hammer the
server but still allows to sync all channels in under a minute.