pkg/syncer/syncer: Extend list of synced channels with LTS
The LTS channels for the different streams should be synced but not the
placeholder LTS channel used to manage the desired version in Nebraska.
Extend teh list of hard-coded synced channels with those that start
with "lts-" denoting an LTS stream channel. Ideally this would be a
boolean field on the channel itself, to set whether to sync or not.