backend: Add and use updateInstanceObjStatus method
The updateInstanceStatus method needs an instanceID and an appID and
then needs to perform a query to get the Instance struct, but there are
places where the caller already has this information, so we save extra
queries if we just use the already known Instance struct. For that, this
patch introduces a new updateInstanceObjStatus method that does exactly
the same as the one mentioned before but using an instance struct.