backend: Don't update instance/instance_application on every ping
Whenever an instance checks for an update we were updating the
database to register the instance and other details
(instance_application); this was being done with a database transaction
since we want to rollback if one of the table upsertions fails. This
means that the whole operation can take longer than a simple write.

This patch optimizes that by:
1. Only using the transaction if the instance *and* instance_application
   entries do not exist;
2. Only updating the instance_application if it's been more than 5
   minutes since the last update (when no other columns need to be
   updated).

This should prevent writes created by constant or abusive update checks
and optimize the writes for existing instances as well.