backend: Simplify and optimize the update granted logic
The update granted logic was updating the respective
instance_application's entry to add the update-granted state and other
data, and then was calling the regular updateInstanceStatus which will
affect the same record. This means there are 2 write calls when we could
use just have one and that impacts performance (besides making the code
more complex).

This patch changes how the updateGranted function works, to reuse the
underlying logic of updateInstanceStatus and thus replacing this
function and garanteeing just a single write for this operation.