backend: Don't register a group activity if there's a recent one
Instead of checking whether there are no updates to the current version
for filing a new group activity ("rollout started"), just check if the
latest one is recent (< 24h). This will allow us to decouple the
mentioned activity from the group's stats which is a complex query and
we only want to use it when really needed.