Add configurable group track names
The mapping of Omaha track to a group ID was hardcoded for the stable,
beta, alpha, and edge channel using only the initial the group IDs.
When the groups were recreated, resolving the track was broken and then
the groups always had to use the group ID as track which is confusing.
For user-created groups there should be a way to add a track name.
Add track name fields to the groups and expose them in the UI instead
of the group ID. The edit and add dialogs allow to enter values but if
omitted, the group ID will be used to populate the field. This is also
done for existing entries through a migration. For requests which refer
a non-existing track, a more clear error code is now returned. The test
data needed some fixes for missing architecture information because the
test requests used the amd64 architecture but the database defaults to
the "all" architecture.