Add cache for group track names and channel architectures
The lookup of the group track name and the channel architecture to find
the group ID has to use an SQL JOIN statement each time which with many
combinations could have a noticable performance impact.
Cache the mapping of group track names and channel architectures to
group IDs in a table and explicitly invalidate it when the group
database entries are touched.