frontend: Replace the Switch component we use for edit/add group dialog
Currently we were using the Switch component from formik-material-ui
which was sending switch values wrapped in array and there was no
way to change this behaviour. We already have material-ui as a
dependency and thus it makes sense to use the Switch component directly
from there and use the onChange handler to set the Switch checked
property on the form field.