cmd/nebraska: Support secret API endpoint suffix for private instances
Private instances that are deployed on the Internet should sometimes
only serve a set of trusted clients. Since the server interacts with
human users via the administration web interface, an IP address filter
on the host system is not convenient.
Allow to move the Omaha API endpoint to a secret path so that only
clients that know the path are able to register. The secret is appended
to the well-known /v1/update path and separated by a slash. For now
there is only one global secret specified as program parameter.
A future feature could be to manage multiple secrets in the web UI
conveniently to handle leaks, transitions, and multiple trustees.