db/migrations: Unify group default settings with public instance
The defaults were varying between the groups and too limiting which
caused confusion when after one request no updates were served anymore.
Also the Safe Mode was set which does not behave as supposed because
when the updates are finally turned off by the Safe Mode they can't be
activated again as the faulty instance is still considered.
Adjust the defaults, affecting only new installations.