backend: Ensure instances are not stuck in an updating state
The Complete event is sometimes ignored similarly to unexpected events
when no update is in progress. This is supposed to prevent Complete
events for already updated instances or those where the previous
version mismatched the version seen by the server. However, the
previous version seen by the server is nothing to be trusted because
the database state may be wrong or from a malicious client. Since
update-engine has inbuilt rollback detection it will create an error
event on its own if it detects that it does not run the new version
instead of a Complete event. The case left to cover is that
update-engine sends Complete events all the time in normal operation
and we want to reset the updating state so that the instance can be
granted updates again. Therefore, do a state transition even though
we drop the event. The Undefined state is chosen because the instance
did not tell that it just now updated from another version but sent an
invalid previous version. Also make sure that the UpdateInProgress
actually relates to the instance state by resetting it for the
Undefined state (and OnHold which was forgotten, too).

Fixes https://github.com/kinvolk/nebraska/issues/303