backend: Set Instance Alias from Omaha request
The instance alias can already be set from the UI but should also be
settable through an Omaha request.
Add the alias as additional information besides the ID and update an
existing alias unless the new alias is an empty string. This required
the go-omaha library to be extended. A test case was added, too.