backend: Don't fetch group stats in events unless it's really needed
The group's stats take a while to fetch, and they were being fetched
only to be used solely if the "disableUpdatesOnFailedRollout" is true.
So the group's stats should rather be gathered only if the mentioned
setting is true which is much quicker to check, and moreover, this
setting is only true during the tests.

This patch thus switches the order of fetching the group's stats +
checking the mentioned setting, which should make things faster.