frontend: Simplify use of selected interval in TimeIntervalLinks
The mentioned component was using the selectedInterval property as an
object containing a displayValue and a queryValue, and erroneously using
the former for comparison. These values are an implementation detail of
the TimeIntervalLinks component, so it should use a simple "query value"
for the mentioned property.

This patch adds that change and also updates the uses of the component.