frontend: Add time interval links to the instances' list
The instance list was showing the instances for a time interval selected
in the previous view. However, there was no context about that interval
to the user (except for the URL).

Thus, these changes add the time interval links to the instances list
view, not only providing the needed context but also allowing the user
to select the interval if needed.