build: Set the version from the original source instead of the container
We were getting the container image version with the -dirty suffix
because it'd be generated from running the git command after the source
was copied but ignoring the files unneeded for the build.

In order to avoid this but still use the correct version from the
original source directory (meaning it'll have the -dirty suffix if there
are changes), the Dockerfile now uses a build arg to set the VERSION
env var.