backend: Allow to consider updates successful with plain ResultSuccess
Since parts of Nebraska are very tied to how Flatcar Container Linux
works, a complete event was only getting saved as so for the instance in
the DB if the result was "SuccessReboot", but we want to also consider
the simple "Success" result, as not all update operations from different
apps will require a reboot.

This patch keeps the Flatcar behavior for the Flatcar application though,
in order to ensure consistency in Nebraska's behavior.