backend: Change instance fetch query to a join query
There are several problems attached to how we handle getting instances
right now.
Our current approach:
We fetch instances and instanceApplications and then map values based
on instance ids the problem with this approach is suppose if we want to
bring in filters like sorting or searching we will have to look what
table does the filter belongs to.
The correct way should be to use a join query to combine instance and
instance app table so that we have everything at one place and thus
applying filters become much easier.
This patch also removes some of the non-required code that is resultant
of the join query conversion.