backend: Change how the groups' stats are calculated
The groups' stats were being computed without taking into account the
version of the group, so the information could be mistakenly
interpreted by users (e.g. having instances in an "update complete"
state but that being for an older version because the instances stopped
querying for updates for a while).

So this patch changes the query that performs the stats in order to show
exactly how many instances are up to date or updating to the current
version.