backend: Count up to date instances without history as complete
New instances that are up to date will not perform updates and thus
have no status related to the update. This would make them show up
in the stats as "undefined" but they should instead be counted as up
to date.

Maybe a distinction between "up to date" and "complete" (went through
an update and has now finished) could be done, but it's at least an
improvement to show those instances as completed.