backend: Allow the use of a product-id in applications
A product-id in the form of an app name including a reverse domain is
allowed since omaha version 3, and it's a human-readable name, making
the use of this ID much easier than the UUID based one.

This patch allows both the new product-id and the UUID to be used when
identifying applications.