backend: Allow to override the packages URL in the syncer
This patch adds a new CLI option to override what the URL should be
for the packages that the syncer creates.

Previously this was only possible with the packagfes host option, but
even then it would be automtatically generated from the Nebraska URL
option. Thus, with this option we give a lot more flexibility.