backend: Add migration for autovac threshold config
Because autovacuum and autoanalyze are effectively disabled when tables
are very large. This is because the default is 20% of a table (and 10%
of a table for analyze).

Instead we change it to when about 5,000 rows change. This value was chosen
based on getting the autovacuum to run every day. Because it's large enough
to not cause the autovac to run all the time, but about the right size
to make a difference for query statistics and reducing table bloat.

The analyze threshold was chosen at half the autovacuum threshold
because the defaults are set at half.