chart: enable secrets injection through secrets operator
Previously, the chart only took a key value map for data that is commonly treated as secret. This patch creates a mechanism that allows users to bring these secrets in with a secrets operator or manually, and define them, setting the relevant environment variables to the value of their secret. This requires a bypass toggle since the charts secret currently requires the secret to be passed in through values.