chart: uncouple nebraska-url from host packages
Previously the chart would only set the -nebraska-url flag if the hostPackages.enable toggle was set to true, but this breaks the chart for any users who aren't looking to set this. If the chart is deployed to a kubernetes cluster for example, the default, localhost, will not work for them. This should be provided by default if the ingress exists.