docs: update Flatcar usage section
Nowadays Nebraska supports group aliases and the UUID is not needed
anymore. While the whitespace in the machine alias works for
update_engine there are some other parts in Flatcar that try to source
the update.conf file and fail when it has unqouted whitespace. Also,
the restart of update-engine.service is not really needed and it's
probably good to point the user to the Flatcar docs for more info.