omaha: Use both SHA1 and SHA256 hashes for extra files
For extra files there is only one hash attribute but it was used for
both a SHA1 and SHA256 entry in the Omaha response. On the wire and in
the UI it's unclear which of the formats it should be.
Use two entries in the database for each hash format and make clear what
the format looks like, i.e., base64 or hex. With a change in go-omaha
we can also omit the empty regular SHA1 hash entry but in any case the
client would be expected to know about the SHA256 attribute and use this
instead. This would match what's done in
https://chromium.googlesource.com/chromium/src.git/+/master/docs/updater/protocol_3_1.md