github: Update ubuntu to LTS and drop older nodejs versions
Drop node 14.x support because it does not support modern
package-lock.json lockfile versions. Node 16, and 18 support the
latest version 3.

Supporting node 14.x is too much maint work for not much gain.

18.x has an incompatibility with the webpackge build tool
currently used by react-scripts we are using. So for now
do not upgrade to use that version of node.

Signed-off-by: Vincent Batts <vincentbatts@microsoft.com>