backend: Add instance statistics table
A by-product of the Omaha update protocol employed by Nebraska delivers statistics on
version spread, but accumulating and presenting this data suffers from a number of
limitations:

- The chart only represents all nodes’ current OS versions. We currently do not have
  historic data of the nodes’ previous software versions, and we would like to track of
  this data.
- While historic data can be calculated from update events, this is computationally
  intense and unfeasible in practice. We currently cannot calculate or store this
  historic data efficiently, and we would like to be able to have quick and easy access
  to this data.
- To be able to back-fill the live Nebraska server using historic data available in CSV
  format, we need to use a third-party tool. We currently cannot accomplish this using
  our existing setup, and we would like to implement a mechanism to do this natively.

We implement the `instance_stats` table (based on instance fields from
`instance_application`) to store data on current version spread in an efficient format
optimized for querying. To accomplish this, there are five data points that we
associate each instance with, serving as the main fields of this table:

- Timestamp: Necessary for keeping track of when update checks are performed
- Channel name: Necessary to identify an individual instance
- Architecture: Necessary to identify an individual instance
- Version: Necessary to identify an individual instance
- Instance count: Necessary for grouping instances by the previous four fields

New entries to the table will be generated using a background job scheduled
periodically at a specified interval.