backend: Use UTC for instance stats default time
The time.Now() function uses the local time which causes a mismatch
and no entries get created when, e.g., being two hours ahead because
the instance would be checking in at 10 UTC but it gets compared to
12 local time and in this time span nothing is able to check in because
the local time will always be too far ahead.
Use the right time zone for the default time of the instance stats
query.