.github: Add GitHub Action to release container image
When creating a release tag we want the image to be built and published
as flatcar/nebraska with the version tag and also as 'latest' image when
the tag is not the special 'staging' tag.
The nebraska-helm chart is also released on the same repo, and thus we
need to filter for strict semver tags.