frontend/Makefile: Add dependencies for npm install
The node_modules folder needs to be updated when the package.json file
changes. This was not encoded as dependency and the rule wasn't "phony"
either, so that we get build errors each time the package.json changes.
Add the package.json file and the lock file as dependencies for the rule
so that when the files are newer than the node_modules folder, we run
"npm install" automatically. We could also use a "phony" rule but since
the build is already slow we can skip the npm invocation if it doesn't
need to run.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>